+++ src/app/dashboard/components/appointment-details/appointment-details.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, EventEmitter, Output } from '@angular/core';
+import { Component, Input, EventEmitter, Output, ViewChild, } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import * as screenfull from 'screenfull';
 
@@ -8,9 +8,10 @@ import { VibiioUpdateService } from '../../services/vibiio-update.service';
 import { SidebarCustomerStatusSharedService } from '../../services/sidebar-customer-status-shared.service';
 import { AvailabilitySharedService } from '../../services/availability-shared.service';
 import { DateFormatService } from '../../../services/date-format.service';
+import { AppointmentDetailsFormStatusService } from '../../services/appointment-details-form-status.service';
 
 // Components
-import { NotesComponent } from '../../containers/notes/notes.component';
+import { ConsumerAddressComponent } from '../consumer-address/consumer-address.component';
 
 // Interfaces
 import { Appointment } from '../../models/appointment.interface';
@@ -19,55 +20,45 @@ import { Vibiio } from '../../models/vibiio.interface';
 import { Note } from '../../models/consumer-note.interface';
 import { ResponseErrorService } from '../../../services/response-error.service';
 import { InsurancePolicy } from '../../models/insurance-policy.interface';
+import { ConsumerUpdateService } from '../../services/consumer-update.service';
+import { Address } from '../../models/address.interface';
 
 @Component({
-    selector: 'appointment-details',
+    selector: 'vib-appointment-details',
     templateUrl: 'appointment-details.component.html',
     styleUrls: ['appointment-details.component.scss']
 })
 
-export class AppointmentDetailsComponent  {
+export class AppointmentDetailsComponent {
+    @Input() vibiioConnecting: boolean;
+    @Input() onVibiio: boolean;
+    @Input() appointment: Appointment;
+    @Input() user: User;
+    @Input() vibiio: Vibiio;
+    @Input() neworkDisconnected: boolean;
+    @Input() timeZone: string;
+    @Input() address: Address;
+
     imgData: string;
     vibiioFullscreen = false;
+    isEditingForms = false;
+    isUpdatingForms = false;
 
-    @Input()
-    vibiioConnecting: boolean;
+    @Output() startVibiio = new EventEmitter<boolean>();
+    @Output() endVibiio = new EventEmitter<boolean>();
+    @Output() claimVibiio: EventEmitter<boolean> = new EventEmitter<boolean>();
+    @Output() refreshNotes: EventEmitter<any> = new EventEmitter<any>();
 
-    @Input()
-    onVibiio: boolean;
-
-    @Input()
-    appointment: Appointment;
-
-    @Input()
-    user: User;
-
-    @Input()
-    vibiio: Vibiio;
-
-    @Input()
-    neworkDisconnected: boolean;
-
-    @Input()
-    timeZone: string;
-
-    @Output()
-    startVibiio: EventEmitter<any> = new EventEmitter<any>();
-
-    @Output()
-    endVibiio: EventEmitter<any> = new EventEmitter<any>();
-
-    @Output()
-    claimVibiio: EventEmitter<boolean> = new EventEmitter<boolean>();
-
-    @Output()
-    refreshNotes: EventEmitter<any> = new EventEmitter<any>();
+    @ViewChild(ConsumerAddressComponent) addressForm: ConsumerAddressComponent;
 
     constructor(private StatusUpdateService: VibiioUpdateService,
                 private sidebarCustomerStatusSharedService: SidebarCustomerStatusSharedService,
                 private availabilitySharedService: AvailabilitySharedService,
                 private dateFormatService: DateFormatService,
-                private router: Router) {}
+                private router: Router,
+                private formStatusService: AppointmentDetailsFormStatusService,
+                private consumerUpdateService:  ConsumerUpdateService) {}
+
 
     updateStatus(event) {
       const options = { status: event.status };
@@ -81,8 +72,18 @@ export class AppointmentDetailsComponent  {
         });
     }
 
+    updateAddress() {
+      if (this.addressForm.editForm.valid) {
+        const address = this.addressForm.editForm.value;
+        this.consumerUpdateService.updateAddress(address)
+          .subscribe( (data) => {
+            this.address = data.address;
+          });
+      }
+    }
+
     connect() {
-      this.startVibiio.emit(event);
+      this.startVibiio.emit(true);
       this.vibiioConnecting = true;
       this.onVibiio = true;
       // check to see if appointment has been claimed and auto assign
@@ -92,7 +93,7 @@ export class AppointmentDetailsComponent  {
     }
 
     disconnect() {
-      this.endVibiio.emit(event);
+      this.endVibiio.emit(true);
       this.vibiioConnecting = false;
       this.availabilitySharedService.emitChange(true);
       this.router.navigateByUrl('/dashboard/vibiio-profile/' + this.vibiio.id);
@@ -110,12 +111,32 @@ export class AppointmentDetailsComponent  {
       return this.dateFormatService.parseTime(time, this.timeZone);
     }
 
-    toggleVibiioFullscreen() {
-      this.vibiioFullscreen = !this.vibiioFullscreen;
+    onEdit(event) {
+      this.isEditingForms = event;
+      this.formStatusService.onFormEdit();
+  }
 
-      // Toggles fullscreen using screenfull package
-      if (screenfull.enabled) {
-        screenfull.toggle();
-      }
+  onUpdate() {
+      this.isUpdatingForms = true;
+      this.formStatusService.onFormUpdate();
+      this.updateAddress();
+  }
+
+  onCancel() {
+      this.formStatusService.onCancel();
+      this.consumerUpdateService.refreshAddress(this.address.id)
+        .subscribe( (data) => {
+          this.addressForm.editForm.patchValue(data.address);
+          this.isEditingForms = false;
+        });
+  }
+
+  toggleVibiioFullscreen() {
+    this.vibiioFullscreen = !this.vibiioFullscreen;
+
+    // Toggles fullscreen using screenfull package
+    if (screenfull.enabled) {
+      screenfull.toggle();
     }
+  }
 }