+++ src/app/dashboard/containers/vibiio-profile/vibiio-profile.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, ViewChild } from '@angular/core';
 import { Routes, RouterModule, Router, ActivatedRoute } from '@angular/router';
 
 // Services
@@ -9,9 +9,15 @@ import { ConsumerProfile } from '../../models/consumer-profile.interface';
 import { VideoSnapshot } from '../../models/video-snapshot.interface';
 import { Note } from '../../models/consumer-note.interface';
 import { Vibiio } from '../../models/vibiio.interface';
+import { InsuranceStatusService } from '../../services/insurance-status.service';
+import { ISubscription } from 'rxjs/Subscription';
+import { OnDestroy } from '@angular/core/src/metadata/lifecycle_hooks';
+import { VibiioProfileFormStatusService } from '../../services/vibiio-profile-form-status.service';
+import { ConsumerUpdateService } from '../../services/consumer-update.service';
+import { ConsumerProfileComponent } from '../../components/consumer-profile/consumer-profile.component';
 
 @Component({
-    selector: 'vibiio-profile',
+    selector: 'vib-vibiio-profile',
     templateUrl: 'vibiio-profile.component.html',
     styleUrls: ['vibiio-profile.component.scss']
 })
@@ -21,9 +27,14 @@ export class VibiioProfileComponent implements OnInit {
     notes: Note[];
     vibiioId: number;
     description: string;
+    isEditing = false;
+    isUpdating = false;
+
+    @ViewChild(ConsumerProfileComponent) consumerProfileChild: ConsumerProfileComponent;
 
     constructor(private activatedRoute: ActivatedRoute,
-                private vibiioProfileService: VibiioProfileService) { }
+                private vibiioProfileService: VibiioProfileService,
+                private formStatusService: VibiioProfileFormStatusService) { }
 
     ngOnInit() {
         this.activatedRoute.data.subscribe( (data) => {
@@ -38,5 +49,23 @@ export class VibiioProfileComponent implements OnInit {
         this.vibiioProfileService.getVibiio(consumerProfileId).subscribe( (data) => {
             this.notes = data.vibiio.notes;
         });
+    }
+
+    onFormEdit(event) {
+        this.isEditing = event;
+    }
+
+    onFormUpdate() {
+        this.isUpdating = true;
+        this.formStatusService.onFormUpdate();
+        this.consumerProfileChild.updateAddress();
+        this.isEditing = false;
+    }
+
+    onCancel() {
+        this.formStatusService.onCancel();
+        this.consumerProfileChild.refreshAddress();
+        this.isEditing = false;
+        this.isUpdating = false;
     }
 }