+++ src/app/dashboard/services/consumer-sort.service.ts
@@ -3,20 +3,14 @@ import { SortType } from '../models/sort-type.interface';
 
 @Injectable()
 export class ConsumerSortService {
-  nameAsc: SortType = { name: 'A - Z', options: { property: 'last_name', desc: true } };
-  nameDesc: SortType =  { name: 'Z - A', options: { property: 'last_name', desc: false } };
-  dateAsc: SortType = { name: 'SOONEST', options: { property: 'appointment_scheduled_datetime', desc: true } };
-  dateDesc: SortType  = { name: 'LATEST', options: { property: 'appointment_scheduled_datetime', desc: false } };
-  status: SortType = { name: 'STATUS', options: { property: 'status', desc: false } };
-  bundle: Array<SortType> = [];
+    nameAsc: SortType = { name: 'A - Z', options: { property: 'last_name', desc: true } };
+    nameDesc: SortType =  { name: 'Z - A', options: { property: 'last_name', desc: false } };
+    dateAsc: SortType = { name: 'SOONEST', options: { property: 'appointment_scheduled_datetime', desc: false } };
+    dateDesc: SortType  = { name: 'LATEST', options: { property: 'appointment_scheduled_datetime', desc: true } };
+    status: SortType = { name: 'STATUS', options: { property: 'status', desc: false } };
+    bundle: Array<SortType> = [this.nameAsc, this.nameDesc, this.dateAsc, this.dateDesc, this.status];
 
   build() {
-    this.addItem(this.nameAsc);
-    this.addItem(this.nameDesc);
-    this.addItem(this.dateAsc);
-    this.addItem(this.dateDesc);
-    this.addItem(this.status);
-
     return this.bundle;
   }