+++ CHANGELOG.md
@@ -1,4 +1,25 @@
-## Vibiio API 0.0.0 (April 7, 2017) ##
+# Vibiio Web Client Changelog
+All notable changes to this project will be documented in this file.
 
-* Authenticate with email and password
-* Integrate tokbox client SKD and setup basic video chat connection
+The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## [2.0.2] - 2017-12-6
+## Added
+- Dummy component to display consumer info on the 'All Consumer' and 'Consumers by Status' views
+
+## Changed
+- Form UX
+
+## Removed
+- edit button from UI
+
+## [2.0.1] - 2017-11-28
+## Fixed
+- Vibiiographer profile display bugs
+
+## [2.0.0] - 2017-11-28
+### Added
+- Start Versioning